Small textual suggestion
As all other items of this manual are to be done in Github, I doubt if it is clear for everyone that this terminal is in RStudio.  (I remember I changed something similar before, is it possible that this text is in 2 places?  Then it might be a good idea to check the earlier changes.)